refactor cli modules and close remaining PR feedback
