+++ crates/uc-cli/src/main.rs
@@ -37,6 +37,12 @@ use walkdir::WalkDir;
 
 #[cfg(feature = "dev-benchmark-command")]
 mod benchmark_cmd;
+mod cache;
+mod daemon;
+mod fingerprint;
+use cache::*;
+use daemon::*;
+use fingerprint::*;
 
 const BUILD_CACHE_SCHEMA_VERSION: u32 = 1;
 const MIN_HASH_LEN: usize = 2;
@@ -1954,1125 +1960,6 @@ fn run_build_with_uc_cache(
     Ok((run, false, fingerprint, telemetry))
 }
 
-fn async_persist_scope_key(workspace_root: &Path, profile: &str) -> String {
-    format!("{}::{profile}", workspace_root.display())
-}
-
-fn async_persist_in_flight_set() -> &'static Mutex<HashSet<String>> {
-    static IN_FLIGHT: OnceLock<Mutex<HashSet<String>>> = OnceLock::new();
-    IN_FLIGHT.get_or_init(|| Mutex::new(HashSet::new()))
-}
-
-fn try_mark_async_persist_in_flight(scope_key: &str) -> bool {
-    let mut in_flight = async_persist_in_flight_set()
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    in_flight.insert(scope_key.to_string())
-}
-
-fn clear_async_persist_in_flight(scope_key: &str) {
-    let mut in_flight = async_persist_in_flight_set()
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    in_flight.remove(scope_key);
-}
-
-fn async_persist_error_slot() -> &'static Mutex<VecDeque<String>> {
-    static SLOT: OnceLock<Mutex<VecDeque<String>>> = OnceLock::new();
-    SLOT.get_or_init(|| Mutex::new(VecDeque::new()))
-}
-
-fn record_async_persist_error(error: String) {
-    let mut slot = async_persist_error_slot()
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    slot.push_back(error);
-    while slot.len() > ASYNC_PERSIST_ERROR_QUEUE_LIMIT {
-        if let Some(dropped) = slot.pop_front() {
-            tracing::warn!(
-                dropped_error = %dropped,
-                "async cache persistence error queue dropped oldest entry"
-            );
-            eprintln!(
-                "uc: warning: async cache persistence error queue dropped oldest entry: {dropped}"
-            );
-        }
-    }
-}
-
-fn take_async_persist_errors() -> Vec<String> {
-    let mut slot = async_persist_error_slot()
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    slot.drain(..).collect()
-}
-
-struct AsyncPersistGuard {
-    scope_key: String,
-}
-
-impl AsyncPersistGuard {
-    fn new(scope_key: String) -> Self {
-        Self { scope_key }
-    }
-}
-
-impl Drop for AsyncPersistGuard {
-    fn drop(&mut self) {
-        clear_async_persist_in_flight(&self.scope_key);
-    }
-}
-
-struct AsyncPersistTask {
-    scope_key: String,
-    workspace_root: PathBuf,
-    profile: String,
-    fingerprint: String,
-    cache_root: PathBuf,
-    objects_dir: PathBuf,
-    entry_path: PathBuf,
-}
-
-fn async_persist_sender() -> &'static SyncSender<AsyncPersistTask> {
-    static SENDER: OnceLock<SyncSender<AsyncPersistTask>> = OnceLock::new();
-    SENDER.get_or_init(|| {
-        let (sender, receiver) = mpsc::sync_channel(ASYNC_PERSIST_QUEUE_LIMIT);
-        thread::spawn(move || run_async_persist_worker(receiver));
-        sender
-    })
-}
-
-fn run_async_persist_worker(receiver: Receiver<AsyncPersistTask>) {
-    for task in receiver {
-        let _guard = AsyncPersistGuard::new(task.scope_key.clone());
-        if let Err(err) = persist_cache_entry_for_build(
-            &task.workspace_root,
-            &task.profile,
-            &task.fingerprint,
-            &task.cache_root,
-            &task.objects_dir,
-            &task.entry_path,
-        ) {
-            let _ = fs::remove_file(&task.entry_path);
-            record_async_persist_error(err.to_string());
-            tracing::warn!(error = %format!("{err:#}"), "async cache persistence failed");
-            eprintln!("uc: warning: async cache persistence failed: {err:#}");
-        }
-    }
-}
-
-fn persist_cache_entry(
-    profile: &str,
-    fingerprint: &str,
-    cached_artifacts: &[CachedArtifact],
-    entry_path: &Path,
-) -> Result<()> {
-    let entry = BuildCacheEntry {
-        schema_version: BUILD_CACHE_SCHEMA_VERSION,
-        fingerprint: fingerprint.to_string(),
-        profile: profile.to_string(),
-        artifacts: cached_artifacts.to_vec(),
-    };
-
-    let bytes = serde_json::to_vec(&entry)?;
-    atomic_write_bytes(entry_path, &bytes, "cache entry")?;
-
-    Ok(())
-}
-
-fn persist_cache_entry_for_build(
-    workspace_root: &Path,
-    profile: &str,
-    fingerprint: &str,
-    cache_root: &Path,
-    objects_dir: &Path,
-    entry_path: &Path,
-) -> Result<()> {
-    let cached_artifacts =
-        collect_cached_artifacts_for_entry(workspace_root, profile, cache_root, objects_dir)?;
-    let _cache_lock = acquire_cache_lock(cache_root)?;
-    persist_cache_entry(profile, fingerprint, &cached_artifacts, entry_path)?;
-    if should_enforce_cache_size_budget_now() {
-        enforce_cache_size_budget(cache_root)?;
-    }
-    Ok(())
-}
-
-fn enforce_cache_size_budget(cache_root: &Path) -> Result<()> {
-    let budget = max_cache_bytes();
-    if budget == 0 || !cache_root.exists() {
-        return Ok(());
-    }
-
-    #[derive(Clone)]
-    struct CacheFile {
-        path: PathBuf,
-        size: u64,
-        modified_ms: u64,
-        is_object: bool,
-    }
-
-    let mut files = Vec::<CacheFile>::new();
-    let mut total = 0_u64;
-    for entry in WalkDir::new(cache_root).follow_links(false).into_iter() {
-        let entry = entry.with_context(|| {
-            format!(
-                "failed to traverse cache tree under {}",
-                cache_root.display()
-            )
-        })?;
-        if !entry.file_type().is_file() {
-            continue;
-        }
-        let path = entry.path();
-        if !is_removable_cache_file(path) {
-            continue;
-        }
-        let metadata =
-            fs::metadata(path).with_context(|| format!("failed to stat {}", path.display()))?;
-        let size = metadata.len();
-        total = total.saturating_add(size);
-        let modified_ms = metadata_modified_unix_ms(&metadata).unwrap_or_default();
-        let is_object = path
-            .components()
-            .any(|c| matches!(c, Component::Normal(name) if name == "objects"));
-        files.push(CacheFile {
-            path: path.to_path_buf(),
-            size,
-            modified_ms,
-            is_object,
-        });
-    }
-
-    if total <= budget {
-        if total > (budget.saturating_mul(9) / 10) {
-            eprintln!("uc: cache usage is high: {total} / {budget} bytes");
-        }
-        return Ok(());
-    }
-
-    files.sort_by(|a, b| {
-        a.is_object
-            .cmp(&b.is_object)
-            .reverse()
-            .then_with(|| a.modified_ms.cmp(&b.modified_ms))
-            .then_with(|| a.path.cmp(&b.path))
-    });
-
-    let mut removed = 0_u64;
-    for file in files {
-        if total <= budget {
-            break;
-        }
-        match fs::remove_file(&file.path) {
-            Ok(()) => {
-                total = total.saturating_sub(file.size);
-                removed = removed.saturating_add(file.size);
-            }
-            Err(err) if err.kind() == io::ErrorKind::NotFound => {
-                total = total.saturating_sub(file.size);
-            }
-            Err(err) => {
-                eprintln!(
-                    "uc: warning: failed to evict cache file {}: {err}",
-                    file.path.display()
-                );
-            }
-        }
-    }
-
-    if removed > 0 {
-        eprintln!(
-            "uc: cache eviction removed {} bytes (budget {} bytes)",
-            removed, budget
-        );
-    }
-    if total > budget {
-        eprintln!(
-            "uc: warning: cache remains over budget after eviction ({} > {} bytes)",
-            total, budget
-        );
-    }
-    Ok(())
-}
-
-fn is_removable_cache_file(path: &Path) -> bool {
-    let Some(name) = path.file_name().and_then(|v| v.to_str()) else {
-        return false;
-    };
-    name != ".lock"
-}
-
-fn load_cache_entry(path: &Path) -> Result<Option<BuildCacheEntry>> {
-    if !path.exists() {
-        return Ok(None);
-    }
-
-    let metadata =
-        fs::metadata(path).with_context(|| format!("failed to stat {}", path.display()))?;
-    let max_bytes = max_cache_entry_bytes();
-    if metadata.len() > max_bytes {
-        eprintln!(
-            "uc: warning: ignoring oversized cache entry {} ({} bytes > {} bytes)",
-            path.display(),
-            metadata.len(),
-            max_bytes
-        );
-        return Ok(None);
-    }
-    let file = File::open(path).with_context(|| format!("failed to read {}", path.display()))?;
-    let mut reader = BufReader::new(file).take(max_bytes + 1);
-    let mut bytes = Vec::new();
-    reader
-        .read_to_end(&mut bytes)
-        .with_context(|| format!("failed to read {}", path.display()))?;
-    if bytes.len() as u64 > max_bytes {
-        eprintln!(
-            "uc: warning: ignoring oversized cache entry {} (>{} bytes)",
-            path.display(),
-            max_bytes
-        );
-        return Ok(None);
-    }
-    let parsed: BuildCacheEntry = match serde_json::from_slice(&bytes) {
-        Ok(entry) => entry,
-        Err(err) => {
-            eprintln!(
-                "uc: warning: ignoring unreadable cache entry {}: {}",
-                path.display(),
-                err
-            );
-            return Ok(None);
-        }
-    };
-    Ok(Some(parsed))
-}
-
-fn artifact_index_entry_matches_expected(
-    index_entry: &ArtifactIndexEntry,
-    metadata: &fs::Metadata,
-    expected_hash: &str,
-    expected_size: u64,
-) -> Result<bool> {
-    let modified_unix_ms = metadata_modified_unix_ms(metadata)?;
-    Ok(index_entry.size_bytes == expected_size
-        && index_entry.size_bytes == metadata.len()
-        && index_entry.modified_unix_ms == modified_unix_ms
-        && index_entry.blake3_hex == expected_hash)
-}
-
-fn upsert_artifact_index_entry_from_metadata(
-    index: &mut ArtifactIndex,
-    relative_path: &str,
-    metadata: &fs::Metadata,
-    expected_hash: &str,
-) -> Result<()> {
-    let modified_unix_ms = metadata_modified_unix_ms(metadata)?;
-    index.entries.insert(
-        relative_path.to_string(),
-        ArtifactIndexEntry {
-            size_bytes: metadata.len(),
-            modified_unix_ms,
-            blake3_hex: expected_hash.to_string(),
-        },
-    );
-    Ok(())
-}
-
-fn restore_cached_artifacts(
-    workspace_root: &Path,
-    profile: &str,
-    cache_root: &Path,
-    objects_dir: &Path,
-    artifacts: &[CachedArtifact],
-) -> Result<bool> {
-    if artifacts.is_empty() {
-        return Ok(false);
-    }
-
-    let index_path = cache_root.join("artifact-index-v1.json");
-    let mut artifact_index = load_artifact_index(&index_path)?;
-    let mut artifact_index_changed = false;
-    let target_root = workspace_root.join("target").join(profile);
-
-    for artifact in artifacts {
-        validate_hex_digest(
-            "cached artifact blake3 hash",
-            &artifact.blake3_hex,
-            MIN_HASH_LEN,
-        )?;
-        validate_cache_object_rel_path(&artifact.object_rel_path)?;
-        let expected_hash = artifact.blake3_hex.as_str();
-        let relative_path = validated_relative_artifact_path(&artifact.relative_path)?;
-        let relative_path_key = normalize_fingerprint_path(&relative_path);
-        let out_path = target_root.join(&relative_path);
-        ensure_path_within_root(&target_root, &out_path, "cache restore path")?;
-
-        if let Ok(existing_metadata) = fs::metadata(&out_path) {
-            let mut matches_cached_artifact = false;
-            if existing_metadata.is_file() {
-                if let Some(index_entry) = artifact_index.entries.get(&relative_path_key) {
-                    if artifact_index_entry_matches_expected(
-                        index_entry,
-                        &existing_metadata,
-                        expected_hash,
-                        artifact.size_bytes,
-                    )? {
-                        matches_cached_artifact = true;
-                    }
-                }
-                if !matches_cached_artifact && existing_metadata.len() == artifact.size_bytes {
-                    if hash_file_blake3(&out_path)? == *expected_hash {
-                        upsert_artifact_index_entry_from_metadata(
-                            &mut artifact_index,
-                            &relative_path_key,
-                            &existing_metadata,
-                            expected_hash,
-                        )?;
-                        artifact_index_changed = true;
-                        matches_cached_artifact = true;
-                    }
-                }
-            }
-            if matches_cached_artifact {
-                continue;
-            }
-        }
-
-        let object_path = objects_dir.join(&artifact.object_rel_path);
-        ensure_path_within_root(objects_dir, &object_path, "cache object path")?;
-        if !object_path.exists() {
-            return Ok(false);
-        }
-        if let Some(parent) = out_path.parent() {
-            fs::create_dir_all(parent)
-                .with_context(|| format!("failed to create {}", parent.display()))?;
-        }
-
-        restore_cache_object(&object_path, &out_path).with_context(|| {
-            format!(
-                "failed to restore cache object {} to {}",
-                object_path.display(),
-                out_path.display()
-            )
-        })?;
-        let restored_metadata = fs::metadata(&out_path)
-            .with_context(|| format!("failed to stat {}", out_path.display()))?;
-        upsert_artifact_index_entry_from_metadata(
-            &mut artifact_index,
-            &relative_path_key,
-            &restored_metadata,
-            expected_hash,
-        )?;
-        artifact_index_changed = true;
-    }
-
-    if artifact_index_changed {
-        artifact_index.schema_version = ARTIFACT_INDEX_SCHEMA_VERSION;
-        if let Err(err) = save_artifact_index(&index_path, &artifact_index) {
-            eprintln!(
-                "uc: warning: failed to update artifact index {}: {err:#}",
-                index_path.display()
-            );
-        }
-    }
-
-    Ok(true)
-}
-
-fn try_reflink_file(source: &Path, destination: &Path) -> io::Result<()> {
-    #[cfg(target_os = "linux")]
-    {
-        let source_file = File::open(source)?;
-        let destination_file = OpenOptions::new()
-            .write(true)
-            .create_new(true)
-            .open(destination)?;
-        let result = unsafe {
-            libc::ioctl(
-                destination_file.as_raw_fd(),
-                libc::FICLONE as _,
-                source_file.as_raw_fd(),
-            )
-        };
-        if result == 0 {
-            return Ok(());
-        }
-        return Err(io::Error::last_os_error());
-    }
-
-    #[cfg(target_os = "macos")]
-    {
-        let source_c = CString::new(source.as_os_str().as_bytes())
-            .map_err(|_| io::Error::new(io::ErrorKind::InvalidInput, "source path contains NUL"))?;
-        let destination_c = CString::new(destination.as_os_str().as_bytes()).map_err(|_| {
-            io::Error::new(io::ErrorKind::InvalidInput, "destination path contains NUL")
-        })?;
-        let result = unsafe { libc::clonefile(source_c.as_ptr(), destination_c.as_ptr(), 0) };
-        if result == 0 {
-            return Ok(());
-        }
-        return Err(io::Error::last_os_error());
-    }
-
-    #[cfg(not(any(target_os = "linux", target_os = "macos")))]
-    {
-        let _ = (source, destination);
-        Err(io::Error::new(
-            io::ErrorKind::Unsupported,
-            "reflink is not supported on this platform",
-        ))
-    }
-}
-
-fn restore_cache_object(source: &Path, destination: &Path) -> Result<()> {
-    if destination.exists() {
-        fs::remove_file(destination)
-            .with_context(|| format!("failed to replace {}", destination.display()))?;
-    }
-
-    if let Err(reflink_err) = try_reflink_file(source, destination) {
-        let _ = fs::remove_file(destination);
-        match fs::hard_link(source, destination) {
-            Ok(()) => return Ok(()),
-            Err(link_err) => {
-                fs::copy(source, destination).with_context(|| {
-                    format!(
-                        "failed to copy cache object after reflink ({}) and hard-link ({}) fallbacks: {} -> {}",
-                        reflink_err,
-                        link_err,
-                        source.display(),
-                        destination.display()
-                    )
-                })?;
-            }
-        }
-    }
-    Ok(())
-}
-
-fn hash_file_blake3(path: &Path) -> Result<String> {
-    let metadata =
-        fs::metadata(path).with_context(|| format!("failed to stat {}", path.display()))?;
-    if metadata.len() > MAX_CACHEABLE_ARTIFACT_BYTES {
-        bail!(
-            "file {} exceeds hashing size limit ({} bytes > {} bytes)",
-            path.display(),
-            metadata.len(),
-            MAX_CACHEABLE_ARTIFACT_BYTES
-        );
-    }
-    let file =
-        fs::File::open(path).with_context(|| format!("failed to open {}", path.display()))?;
-    let mut reader = BufReader::new(file);
-    let mut buf = [0_u8; 8192];
-    let mut hasher = Hasher::new();
-
-    loop {
-        let read = reader
-            .read(&mut buf)
-            .with_context(|| format!("failed to read {}", path.display()))?;
-        if read == 0 {
-            break;
-        }
-        hasher.update(&buf[..read]);
-    }
-
-    Ok(hasher.finalize().to_hex().to_string())
-}
-
-fn hash_fingerprint_source_file(path: &Path) -> Result<String> {
-    let is_cairo = path
-        .extension()
-        .and_then(|ext| ext.to_str())
-        .map(|ext| ext.eq_ignore_ascii_case("cairo"))
-        .unwrap_or(false);
-    if is_cairo {
-        return hash_cairo_source_semantic(path);
-    }
-    hash_file_blake3(path)
-}
-
-fn hash_cairo_source_semantic(path: &Path) -> Result<String> {
-    let bytes = fs::read(path).with_context(|| format!("failed to read {}", path.display()))?;
-    let mut normalized = strip_cairo_comments(&bytes);
-    while matches!(normalized.last(), Some(b' ' | b'\t' | b'\r' | b'\n')) {
-        normalized.pop();
-    }
-    let mut hasher = Hasher::new();
-    hasher.update(b"uc-cairo-semantic-hash-v1");
-    hasher.update(&normalized);
-    Ok(hasher.finalize().to_hex().to_string())
-}
-
-fn strip_cairo_comments(input: &[u8]) -> Vec<u8> {
-    #[derive(Clone, Copy)]
-    enum Mode {
-        Code,
-        LineComment,
-        BlockComment { depth: u32 },
-        SingleQuote,
-        DoubleQuote,
-    }
-
-    let mut mode = Mode::Code;
-    let mut out = Vec::with_capacity(input.len());
-    let mut i = 0_usize;
-
-    while i < input.len() {
-        let b = input[i];
-        let next = input.get(i + 1).copied();
-
-        match mode {
-            Mode::Code => {
-                if b == b'/' && next == Some(b'/') {
-                    mode = Mode::LineComment;
-                    i += 2;
-                    continue;
-                }
-                if b == b'/' && next == Some(b'*') {
-                    mode = Mode::BlockComment { depth: 1 };
-                    i += 2;
-                    continue;
-                }
-                out.push(b);
-                if b == b'\'' {
-                    mode = Mode::SingleQuote;
-                } else if b == b'"' {
-                    mode = Mode::DoubleQuote;
-                }
-                i += 1;
-            }
-            Mode::LineComment => {
-                if b == b'\n' {
-                    out.push(b'\n');
-                    mode = Mode::Code;
-                }
-                i += 1;
-            }
-            Mode::BlockComment { depth } => {
-                if b == b'/' && next == Some(b'*') {
-                    mode = Mode::BlockComment { depth: depth + 1 };
-                    i += 2;
-                    continue;
-                }
-                if b == b'*' && next == Some(b'/') {
-                    if depth <= 1 {
-                        mode = Mode::Code;
-                    } else {
-                        mode = Mode::BlockComment { depth: depth - 1 };
-                    }
-                    i += 2;
-                    continue;
-                }
-                if b == b'\n' {
-                    out.push(b'\n');
-                }
-                i += 1;
-            }
-            Mode::SingleQuote => {
-                out.push(b);
-                if b == b'\\' {
-                    if let Some(escaped) = next {
-                        out.push(escaped);
-                        i += 2;
-                        continue;
-                    }
-                }
-                if b == b'\'' {
-                    mode = Mode::Code;
-                }
-                i += 1;
-            }
-            Mode::DoubleQuote => {
-                out.push(b);
-                if b == b'\\' {
-                    if let Some(escaped) = next {
-                        out.push(escaped);
-                        i += 2;
-                        continue;
-                    }
-                }
-                if b == b'"' {
-                    mode = Mode::Code;
-                }
-                i += 1;
-            }
-        }
-    }
-
-    out
-}
-
-fn metadata_modified_unix_ms(metadata: &fs::Metadata) -> Result<u64> {
-    let modified = metadata
-        .modified()
-        .context("failed to read file modification time")?;
-    let since_epoch = modified.duration_since(UNIX_EPOCH).unwrap_or_default();
-    u64::try_from(since_epoch.as_millis()).context("file modified time overflowed u64")
-}
-
-fn normalize_fingerprint_path(path: &Path) -> String {
-    let raw = path.to_string_lossy();
-    let without_windows_prefix = raw.strip_prefix("\\\\?\\").unwrap_or(&raw);
-    without_windows_prefix.replace('\\', "/")
-}
-
-fn atomic_write_bytes(path: &Path, bytes: &[u8], label: &str) -> Result<()> {
-    static NEXT_TEMP_ID: AtomicU64 = AtomicU64::new(1);
-    let parent = path
-        .parent()
-        .context("cannot atomically write file without parent directory")?;
-    fs::create_dir_all(parent).with_context(|| format!("failed to create {}", parent.display()))?;
-    let stem = path.file_name().and_then(|v| v.to_str()).unwrap_or("file");
-    let temp_id = NEXT_TEMP_ID.fetch_add(1, Ordering::Relaxed);
-    let thread_id = format!("{:?}", thread::current().id());
-    let temp_path = parent.join(format!(
-        ".{stem}.tmp.{}.{}.{}.{}",
-        std::process::id(),
-        thread_id,
-        temp_id,
-        epoch_ms_u64().unwrap_or_default()
-    ));
-    fs::write(&temp_path, bytes).with_context(|| {
-        format!(
-            "failed to write temporary {label} file {}",
-            temp_path.display()
-        )
-    })?;
-    if let Err(err) = fs::rename(&temp_path, path) {
-        let _ = fs::remove_file(&temp_path);
-        return Err(err).with_context(|| {
-            format!(
-                "failed to move temporary {label} {} to {}",
-                temp_path.display(),
-                path.display()
-            )
-        });
-    }
-    Ok(())
-}
-
-fn load_fingerprint_index(path: &Path) -> Result<FingerprintIndex> {
-    if !path.exists() {
-        return Ok(FingerprintIndex::empty());
-    }
-    let bytes = read_bytes_with_limit(path, max_fingerprint_index_bytes(), "fingerprint index")?;
-    match serde_json::from_slice::<FingerprintIndex>(&bytes) {
-        Ok(index) if index.schema_version == FINGERPRINT_INDEX_SCHEMA_VERSION => Ok(index),
-        Ok(_) => Ok(FingerprintIndex::empty()),
-        Err(err) => {
-            eprintln!(
-                "uc: warning: ignoring unreadable fingerprint index {}: {}",
-                path.display(),
-                err
-            );
-            Ok(FingerprintIndex::empty())
-        }
-    }
-}
-
-fn save_fingerprint_index(path: &Path, index: &FingerprintIndex) -> Result<()> {
-    let bytes = serde_json::to_vec(index).context("failed to encode fingerprint index")?;
-    atomic_write_bytes(path, &bytes, "fingerprint index")?;
-    Ok(())
-}
-
-fn fingerprint_index_cache() -> &'static Mutex<HashMap<String, FingerprintIndex>> {
-    static CACHE: OnceLock<Mutex<HashMap<String, FingerprintIndex>>> = OnceLock::new();
-    CACHE.get_or_init(|| Mutex::new(HashMap::new()))
-}
-
-fn fingerprint_index_cache_key(path: &Path) -> String {
-    normalize_fingerprint_path(path)
-}
-
-fn load_fingerprint_index_cached(path: &Path) -> Result<FingerprintIndex> {
-    let key = fingerprint_index_cache_key(path);
-    {
-        let cache = fingerprint_index_cache()
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        if let Some(index) = cache.get(&key) {
-            return Ok(index.clone());
-        }
-    }
-
-    let loaded = load_fingerprint_index(path)?;
-    let mut cache = fingerprint_index_cache()
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    cache.insert(key, loaded.clone());
-    Ok(loaded)
-}
-
-fn store_fingerprint_index_cached(path: &Path, index: &FingerprintIndex) {
-    let key = fingerprint_index_cache_key(path);
-    let mut cache = fingerprint_index_cache()
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    cache.insert(key, index.clone());
-}
-
-fn load_artifact_index(path: &Path) -> Result<ArtifactIndex> {
-    if !path.exists() {
-        return Ok(ArtifactIndex::empty());
-    }
-    let bytes = read_bytes_with_limit(path, max_artifact_index_bytes(), "artifact index")?;
-    match serde_json::from_slice::<ArtifactIndex>(&bytes) {
-        Ok(index) if index.schema_version == ARTIFACT_INDEX_SCHEMA_VERSION => Ok(index),
-        Ok(_) => Ok(ArtifactIndex::empty()),
-        Err(err) => {
-            eprintln!(
-                "uc: warning: ignoring unreadable artifact index {}: {}",
-                path.display(),
-                err
-            );
-            Ok(ArtifactIndex::empty())
-        }
-    }
-}
-
-fn save_artifact_index(path: &Path, index: &ArtifactIndex) -> Result<()> {
-    let bytes = serde_json::to_vec(index).context("failed to encode artifact index")?;
-    atomic_write_bytes(path, &bytes, "artifact index")?;
-    Ok(())
-}
-
-fn compute_build_fingerprint(
-    workspace_root: &Path,
-    manifest_path: &Path,
-    common: &BuildCommonArgs,
-    profile: &str,
-    cache_root: Option<&Path>,
-) -> Result<String> {
-    let scarb_version = scarb_version_line()?;
-    compute_build_fingerprint_with_scarb_version(
-        workspace_root,
-        manifest_path,
-        common,
-        profile,
-        cache_root,
-        &scarb_version,
-    )
-}
-
-fn build_fingerprint_context_digest(
-    canonical_manifest: &Path,
-    common: &BuildCommonArgs,
-    profile: &str,
-    scarb_version: &str,
-) -> String {
-    let mut hasher = Hasher::new();
-    hasher.update(b"uc-build-fingerprint-context-v1");
-    hasher.update(scarb_version.as_bytes());
-    hasher.update(compute_build_env_fingerprint().as_bytes());
-    hasher.update(normalize_fingerprint_path(canonical_manifest).as_bytes());
-    hasher.update(profile.as_bytes());
-    hasher.update(common.package.as_deref().unwrap_or("*").as_bytes());
-    hasher.update(if common.workspace {
-        b"workspace"
-    } else {
-        b"package"
-    });
-    hasher.update(if common.offline {
-        b"offline"
-    } else {
-        b"online"
-    });
-    let mut features = common.features.clone();
-    features.sort_unstable();
-    features.dedup();
-    for feature in features {
-        hasher.update(feature.as_bytes());
-        hasher.update(b",");
-    }
-    hasher.finalize().to_hex().to_string()
-}
-
-fn try_reuse_hot_fingerprint(
-    workspace_root: &Path,
-    index: &FingerprintIndex,
-    context_digest: &str,
-    now_ms: u64,
-    mtime_recheck_window_ms: u64,
-) -> Result<Option<String>> {
-    if index.context_digest.as_deref() != Some(context_digest) {
-        return Ok(None);
-    }
-    let Some(last_fingerprint) = index.last_fingerprint.as_ref() else {
-        return Ok(None);
-    };
-    if last_fingerprint.is_empty() {
-        return Ok(None);
-    }
-
-    for (relative_dir, expected_modified_unix_ms) in &index.directories {
-        let dir_path = if relative_dir == "." {
-            workspace_root.to_path_buf()
-        } else {
-            workspace_root.join(relative_dir)
-        };
-        let metadata = match fs::metadata(&dir_path) {
-            Ok(metadata) => metadata,
-            Err(_) => return Ok(None),
-        };
-        if !metadata.is_dir() {
-            return Ok(None);
-        }
-        let modified_unix_ms = metadata_modified_unix_ms(&metadata)?;
-        if modified_unix_ms != *expected_modified_unix_ms {
-            return Ok(None);
-        }
-    }
-
-    for (relative_path, cached_entry) in &index.entries {
-        let file_path = workspace_root.join(relative_path);
-        let metadata = match fs::metadata(&file_path) {
-            Ok(metadata) => metadata,
-            Err(_) => return Ok(None),
-        };
-        if !metadata.is_file() {
-            return Ok(None);
-        }
-        let modified_unix_ms = metadata_modified_unix_ms(&metadata)?;
-        if metadata.len() != cached_entry.size_bytes
-            || modified_unix_ms != cached_entry.modified_unix_ms
-        {
-            return Ok(None);
-        }
-        if now_ms.saturating_sub(modified_unix_ms) <= mtime_recheck_window_ms {
-            return Ok(None);
-        }
-    }
-
-    Ok(Some(last_fingerprint.clone()))
-}
-
-fn track_fingerprint_directories_for_relative_path(
-    tracked_directories: &mut BTreeSet<String>,
-    relative_path: &Path,
-) {
-    tracked_directories.insert(".".to_string());
-    let mut cursor = relative_path.parent();
-    while let Some(parent) = cursor {
-        if parent.as_os_str().is_empty() {
-            break;
-        }
-        tracked_directories.insert(normalize_fingerprint_path(parent));
-        cursor = parent.parent();
-    }
-}
-
-fn snapshot_tracked_fingerprint_directories(
-    workspace_root: &Path,
-    tracked_directories: &BTreeSet<String>,
-) -> Result<BTreeMap<String, u64>> {
-    let mut snapshot = BTreeMap::new();
-    for relative_dir in tracked_directories {
-        let dir_path = if relative_dir == "." {
-            workspace_root.to_path_buf()
-        } else {
-            workspace_root.join(relative_dir)
-        };
-        let metadata = fs::metadata(&dir_path)
-            .with_context(|| format!("failed to stat {}", dir_path.display()))?;
-        if !metadata.is_dir() {
-            continue;
-        }
-        snapshot.insert(relative_dir.clone(), metadata_modified_unix_ms(&metadata)?);
-    }
-    Ok(snapshot)
-}
-
-fn compute_build_fingerprint_with_scarb_version(
-    workspace_root: &Path,
-    manifest_path: &Path,
-    common: &BuildCommonArgs,
-    profile: &str,
-    cache_root: Option<&Path>,
-    scarb_version: &str,
-) -> Result<String> {
-    let canonical_manifest = manifest_path
-        .canonicalize()
-        .with_context(|| format!("failed to canonicalize {}", manifest_path.display()))?;
-    let context_digest =
-        build_fingerprint_context_digest(&canonical_manifest, common, profile, scarb_version);
-
-    let (index_path, mut index) = if let Some(root) = cache_root {
-        let path = root.join("fingerprint/index-v1.json");
-        (Some(path.clone()), load_fingerprint_index_cached(&path)?)
-    } else {
-        (None, FingerprintIndex::empty())
-    };
-    let max_files = max_fingerprint_files();
-    let max_file_bytes = max_fingerprint_file_bytes();
-    let max_total_bytes = max_fingerprint_total_bytes();
-    let fingerprint_timeout = Duration::from_millis(fingerprint_timeout_ms());
-    let fingerprint_started = Instant::now();
-    let mtime_recheck_window_ms = fingerprint_mtime_recheck_window_ms();
-    let now_ms = epoch_ms_u64().unwrap_or_default();
-
-    if let Some(reused) = try_reuse_hot_fingerprint(
-        workspace_root,
-        &index,
-        &context_digest,
-        now_ms,
-        mtime_recheck_window_ms,
-    )? {
-        return Ok(reused);
-    }
-
-    let mut hasher = Hasher::new();
-    hasher.update(b"uc-build-fingerprint-v2");
-    hasher.update(context_digest.as_bytes());
-
-    let mut updated_entries: BTreeMap<String, FingerprintIndexEntry> = BTreeMap::new();
-    let mut tracked_directories: BTreeSet<String> = BTreeSet::from([".".to_string()]);
-
-    let mut files = Vec::new();
-    let walker = WalkDir::new(workspace_root)
-        .follow_links(false)
-        .max_depth(MAX_FINGERPRINT_DEPTH)
-        .into_iter()
-        .filter_entry(|entry| !is_ignored_entry(workspace_root, entry.path()));
-
-    for entry in walker.filter_map(|e| e.ok()) {
-        if fingerprint_started.elapsed() > fingerprint_timeout {
-            bail!(
-                "fingerprinting timed out after {} ms",
-                fingerprint_timeout.as_millis()
-            );
-        }
-        if !entry.file_type().is_file() {
-            continue;
-        }
-        let path = entry.path();
-        if should_include_fingerprint_file(path) {
-            if files.len() >= max_files {
-                bail!(
-                    "workspace has too many fingerprintable files (>{max_files}); refusing to hash more"
-                );
-            }
-            files.push(path.to_path_buf());
-        }
-    }
-    files.sort();
-    let mut total_fingerprint_bytes = 0_u64;
-
-    for path in files {
-        if fingerprint_started.elapsed() > fingerprint_timeout {
-            bail!(
-                "fingerprinting timed out after {} ms",
-                fingerprint_timeout.as_millis()
-            );
-        }
-        let rel = path
-            .strip_prefix(workspace_root)
-            .unwrap_or(&path)
-            .to_string_lossy()
-            .replace('\\', "/");
-        track_fingerprint_directories_for_relative_path(&mut tracked_directories, Path::new(&rel));
-        let metadata =
-            fs::metadata(&path).with_context(|| format!("failed to stat {}", path.display()))?;
-        let file_size = metadata.len();
-        if file_size > max_file_bytes {
-            bail!(
-                "fingerprint file {} exceeds size limit ({} bytes > {} bytes)",
-                path.display(),
-                file_size,
-                max_file_bytes
-            );
-        }
-        total_fingerprint_bytes = total_fingerprint_bytes.saturating_add(file_size);
-        if total_fingerprint_bytes > max_total_bytes {
-            bail!(
-                "fingerprint source budget exceeded ({} bytes > {} bytes)",
-                total_fingerprint_bytes,
-                max_total_bytes
-            );
-        }
-        let modified_unix_ms = metadata_modified_unix_ms(&metadata)?;
-        let file_hash = if let Some(cached) = index.entries.get(&rel) {
-            let should_rehash_recent =
-                now_ms.saturating_sub(modified_unix_ms) <= mtime_recheck_window_ms;
-            if cached.size_bytes == file_size
-                && cached.modified_unix_ms == modified_unix_ms
-                && !should_rehash_recent
-            {
-                cached.blake3_hex.clone()
-            } else {
-                hash_fingerprint_source_file(&path)?
-            }
-        } else {
-            hash_fingerprint_source_file(&path)?
-        };
-        updated_entries.insert(
-            rel.clone(),
-            FingerprintIndexEntry {
-                size_bytes: file_size,
-                modified_unix_ms,
-                blake3_hex: file_hash.clone(),
-            },
-        );
-        hasher.update(rel.as_bytes());
-        hasher.update(b":");
-        hasher.update(file_hash.as_bytes());
-        hasher.update(b"\n");
-    }
-    let tracked_directory_mtimes =
-        snapshot_tracked_fingerprint_directories(workspace_root, &tracked_directories)?;
-    let fingerprint = hasher.finalize().to_hex().to_string();
-    if let Some(path) = index_path {
-        let changed = index.entries != updated_entries
-            || index.directories != tracked_directory_mtimes
-            || index.context_digest.as_deref() != Some(context_digest.as_str())
-            || index.last_fingerprint.as_deref() != Some(fingerprint.as_str())
-            || index.schema_version != FINGERPRINT_INDEX_SCHEMA_VERSION;
-        index.schema_version = FINGERPRINT_INDEX_SCHEMA_VERSION;
-        index.entries = updated_entries;
-        index.directories = tracked_directory_mtimes;
-        index.context_digest = Some(context_digest);
-        index.last_fingerprint = Some(fingerprint.clone());
-        store_fingerprint_index_cached(&path, &index);
-        if changed {
-            if let Err(err) = save_fingerprint_index(&path, &index) {
-                eprintln!(
-                    "uc: warning: failed to update fingerprint index {}: {err:#}",
-                    path.display()
-                );
-            }
-        }
-    }
-
-    Ok(fingerprint)
-}
-
-fn is_ignored_entry(root: &Path, path: &Path) -> bool {
-    if path == root {
-        return false;
-    }
-    let Some(name) = path.file_name().and_then(|s| s.to_str()) else {
-        return false;
-    };
-    matches!(name, ".git" | "target" | ".scarb" | ".uc")
-}
-
-fn should_include_fingerprint_file(path: &Path) -> bool {
-    let Some(name) = path.file_name().and_then(|s| s.to_str()) else {
-        return false;
-    };
-
-    if matches!(name, "Scarb.toml" | "Scarb.lock" | "Uc.toml") {
-        return true;
-    }
-
-    path.extension()
-        .and_then(|s| s.to_str())
-        .map(|ext| ext == "cairo")
-        .unwrap_or(false)
-}
-
 fn scarb_build_command(common: &BuildCommonArgs, manifest_path: &Path) -> (Command, Vec<String>) {
     let mut command = Command::new("scarb");
     let mut command_vec = vec!["scarb".to_string()];
@@ -4048,306 +2935,6 @@ fn is_process_alive(pid: u32) -> bool {
     system.process(target).is_some()
 }
 
-fn daemon_socket_path(override_path: Option<PathBuf>) -> Result<PathBuf> {
-    if let Some(path) = override_path {
-        return Ok(path);
-    }
-    if let Some(path) = std::env::var_os("UC_DAEMON_SOCKET_PATH") {
-        return Ok(PathBuf::from(path));
-    }
-    let home = std::env::var_os("HOME").context("HOME is not set; provide --socket-path")?;
-    Ok(PathBuf::from(home).join(".uc/daemon/uc.sock"))
-}
-
-fn daemon_log_path(socket_path: &Path) -> PathBuf {
-    socket_path.with_extension("log")
-}
-
-fn remove_socket_if_exists(path: &Path) -> Result<()> {
-    match fs::remove_file(path) {
-        Ok(()) => Ok(()),
-        Err(err) if err.kind() == io::ErrorKind::NotFound => Ok(()),
-        Err(err) => Err(err).with_context(|| format!("failed to remove {}", path.display())),
-    }
-}
-
-fn rotate_daemon_log_if_needed(log_path: &Path) -> Result<()> {
-    let metadata = match fs::metadata(log_path) {
-        Ok(metadata) => metadata,
-        Err(err) if err.kind() == io::ErrorKind::NotFound => return Ok(()),
-        Err(err) => {
-            return Err(err).with_context(|| format!("failed to stat {}", log_path.display()));
-        }
-    };
-    if metadata.len() < DAEMON_LOG_ROTATE_BYTES {
-        return Ok(());
-    }
-    let rotated = PathBuf::from(format!("{}.1", log_path.display()));
-    if rotated.exists() {
-        fs::remove_file(&rotated)
-            .with_context(|| format!("failed to remove {}", rotated.display()))?;
-    }
-    fs::rename(log_path, &rotated).with_context(|| {
-        format!(
-            "failed to rotate daemon log {} to {}",
-            log_path.display(),
-            rotated.display()
-        )
-    })?;
-    Ok(())
-}
-
-fn read_line_limited<R: BufRead>(reader: &mut R, max_bytes: usize, label: &str) -> Result<String> {
-    let mut bytes = Vec::with_capacity(128);
-    loop {
-        let mut byte = [0_u8; 1];
-        let read = reader
-            .read(&mut byte)
-            .with_context(|| format!("failed to read {label}"))?;
-        if read == 0 {
-            break;
-        }
-        if byte[0] == b'\n' {
-            break;
-        }
-        bytes.push(byte[0]);
-        if bytes.len() > max_bytes {
-            bail!("{label} exceeds size limit ({max_bytes} bytes)");
-        }
-    }
-    if bytes.is_empty() {
-        return Ok(String::new());
-    }
-    String::from_utf8(bytes).with_context(|| format!("{label} is not valid UTF-8"))
-}
-
-#[cfg(unix)]
-fn daemon_request(socket_path: &Path, request: &DaemonRequest) -> Result<DaemonResponse> {
-    let mut stream = UnixStream::connect(socket_path)
-        .with_context(|| format!("failed to connect daemon socket {}", socket_path.display()))?;
-    stream
-        .set_read_timeout(Some(Duration::from_secs(120)))
-        .with_context(|| format!("failed to set read timeout for {}", socket_path.display()))?;
-    stream
-        .set_write_timeout(Some(Duration::from_secs(30)))
-        .with_context(|| format!("failed to set write timeout for {}", socket_path.display()))?;
-
-    let payload = serde_json::to_vec(request).context("failed to encode daemon request")?;
-    stream
-        .write_all(&payload)
-        .context("failed to write daemon request payload")?;
-    stream
-        .write_all(b"\n")
-        .context("failed to write daemon request newline")?;
-    stream.flush().context("failed to flush daemon request")?;
-
-    let response_line = {
-        let mut reader = BufReader::new(&mut stream);
-        read_line_limited(
-            &mut reader,
-            DAEMON_REQUEST_SIZE_LIMIT_BYTES,
-            "daemon response",
-        )?
-    };
-    if response_line.trim().is_empty() {
-        bail!("daemon returned empty response");
-    }
-    serde_json::from_str(response_line.trim_end()).context("failed to decode daemon response")
-}
-
-#[cfg(unix)]
-fn daemon_ping(socket_path: &Path) -> Result<DaemonStatusPayload> {
-    match daemon_request(socket_path, &DaemonRequest::Ping)? {
-        DaemonResponse::Pong(status) => Ok(status),
-        DaemonResponse::Error { message } => bail!("daemon ping failed: {message}"),
-        _ => bail!("unexpected daemon response to ping"),
-    }
-}
-
-#[cfg(unix)]
-fn daemon_status_snapshot(
-    base: &DaemonStatusPayload,
-    health: &Arc<Mutex<DaemonHealth>>,
-) -> DaemonStatusPayload {
-    let snapshot = health
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner())
-        .clone();
-    DaemonStatusPayload {
-        pid: base.pid,
-        started_at_epoch_ms: base.started_at_epoch_ms,
-        socket_path: base.socket_path.clone(),
-        protocol_version: base.protocol_version.clone(),
-        healthy: snapshot.consecutive_failures < 3,
-        total_requests: snapshot.total_requests,
-        failed_requests: snapshot.failed_requests,
-        rate_limited_requests: snapshot.rate_limited_requests,
-        last_error: snapshot.last_error,
-    }
-}
-
-#[cfg(unix)]
-fn record_daemon_success(health: &Arc<Mutex<DaemonHealth>>) {
-    let mut state = health
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    state.total_requests = state.total_requests.saturating_add(1);
-    state.consecutive_failures = 0;
-    state.last_error = None;
-    state.last_failure_at = None;
-}
-
-#[cfg(unix)]
-fn record_daemon_failure(health: &Arc<Mutex<DaemonHealth>>, error: String) {
-    let mut state = health
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    state.total_requests = state.total_requests.saturating_add(1);
-    state.failed_requests = state.failed_requests.saturating_add(1);
-    state.consecutive_failures = state.consecutive_failures.saturating_add(1);
-    state.last_error = Some(error);
-    state.last_failure_at = Some(Instant::now());
-}
-
-#[cfg(unix)]
-fn record_daemon_rate_limit(health: &Arc<Mutex<DaemonHealth>>) {
-    let mut state = health
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    state.total_requests = state.total_requests.saturating_add(1);
-    state.failed_requests = state.failed_requests.saturating_add(1);
-    state.rate_limited_requests = state.rate_limited_requests.saturating_add(1);
-    state.consecutive_failures = state.consecutive_failures.saturating_add(1);
-    state.last_error = Some("daemon rate limit exceeded; retry shortly".to_string());
-    state.last_failure_at = Some(Instant::now());
-}
-
-#[cfg(unix)]
-fn maybe_auto_recover_daemon_health(health: &Arc<Mutex<DaemonHealth>>) {
-    let mut state = health
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    if state.consecutive_failures < 3 {
-        return;
-    }
-    let Some(last_failure_at) = state.last_failure_at else {
-        return;
-    };
-    if last_failure_at.elapsed() >= Duration::from_secs(DAEMON_UNHEALTHY_RECOVERY_SECONDS) {
-        state.consecutive_failures = 0;
-        state.last_error = None;
-        state.last_failure_at = None;
-    }
-}
-
-#[cfg(unix)]
-fn handle_daemon_connection(
-    mut stream: UnixStream,
-    status: &DaemonStatusPayload,
-    health: &Arc<Mutex<DaemonHealth>>,
-    should_shutdown: &mut bool,
-    rate_limiter: &mut DaemonRateLimiter,
-) -> Result<()> {
-    stream
-        .set_read_timeout(Some(Duration::from_secs(30)))
-        .context("failed to set daemon read timeout")?;
-    stream
-        .set_write_timeout(Some(Duration::from_secs(120)))
-        .context("failed to set daemon write timeout")?;
-
-    let request_line = {
-        let mut reader = BufReader::new(&mut stream);
-        read_line_limited(
-            &mut reader,
-            DAEMON_REQUEST_SIZE_LIMIT_BYTES,
-            "daemon request",
-        )?
-    };
-    if request_line.trim().is_empty() {
-        return Ok(());
-    }
-    maybe_auto_recover_daemon_health(health);
-
-    if !rate_limiter.allow() {
-        record_daemon_rate_limit(health);
-        let response = DaemonResponse::Error {
-            message: "daemon rate limit exceeded; retry shortly".to_string(),
-        };
-        let payload = serde_json::to_vec(&response).context("failed to encode daemon response")?;
-        stream
-            .write_all(&payload)
-            .context("failed to write daemon response")?;
-        stream
-            .write_all(b"\n")
-            .context("failed to write daemon response newline")?;
-        stream.flush().context("failed to flush daemon response")?;
-        return Ok(());
-    }
-
-    let request: DaemonRequest = match serde_json::from_str(request_line.trim_end()) {
-        Ok(request) => request,
-        Err(err) => {
-            let message = format!("failed to parse daemon request: {err}");
-            record_daemon_failure(health, message.clone());
-            let response = DaemonResponse::Error { message };
-            let payload =
-                serde_json::to_vec(&response).context("failed to encode daemon response")?;
-            stream
-                .write_all(&payload)
-                .context("failed to write daemon response")?;
-            stream
-                .write_all(b"\n")
-                .context("failed to write daemon response newline")?;
-            stream.flush().context("failed to flush daemon response")?;
-            return Ok(());
-        }
-    };
-
-    let response = match request {
-        DaemonRequest::Ping => {
-            record_daemon_success(health);
-            DaemonResponse::Pong(daemon_status_snapshot(status, health))
-        }
-        DaemonRequest::Shutdown => {
-            record_daemon_success(health);
-            *should_shutdown = true;
-            DaemonResponse::Ack
-        }
-        DaemonRequest::Build(request) => match execute_daemon_build(request) {
-            Ok(result) => {
-                record_daemon_success(health);
-                DaemonResponse::Build(result)
-            }
-            Err(err) => {
-                let message = format!("{err:#}");
-                record_daemon_failure(health, message.clone());
-                DaemonResponse::Error { message }
-            }
-        },
-        DaemonRequest::Metadata(request) => match execute_daemon_metadata(request) {
-            Ok(result) => {
-                record_daemon_success(health);
-                DaemonResponse::Metadata(result)
-            }
-            Err(err) => {
-                let message = format!("{err:#}");
-                record_daemon_failure(health, message.clone());
-                DaemonResponse::Error { message }
-            }
-        },
-    };
-
-    let payload = serde_json::to_vec(&response).context("failed to encode daemon response")?;
-    stream
-        .write_all(&payload)
-        .context("failed to write daemon response")?;
-    stream
-        .write_all(b"\n")
-        .context("failed to write daemon response newline")?;
-    stream.flush().context("failed to flush daemon response")?;
-    Ok(())
-}
-
 fn read_bytes_with_limit(path: &Path, max_bytes: u64, label: &str) -> Result<Vec<u8>> {
     let metadata =
         fs::metadata(path).with_context(|| format!("failed to stat {}", path.display()))?;
@@ -4530,1029 +3117,4 @@ fn validate_scarb_toolchain() -> Result<()> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs;
-    use std::sync::{Mutex, OnceLock as TestOnceLock};
-    use std::thread;
-
-    fn unique_test_dir(prefix: &str) -> PathBuf {
-        let nanos = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("system clock before UNIX_EPOCH")
-            .as_nanos();
-        let dir = std::env::temp_dir().join(format!("{prefix}-{}-{nanos}", std::process::id()));
-        fs::create_dir_all(&dir).expect("failed to create test directory");
-        dir
-    }
-
-    fn integration_env_lock() -> &'static Mutex<()> {
-        static LOCK: TestOnceLock<Mutex<()>> = TestOnceLock::new();
-        LOCK.get_or_init(|| Mutex::new(()))
-    }
-
-    fn scarb_available() -> bool {
-        Command::new("scarb")
-            .arg("--version")
-            .output()
-            .map(|output| output.status.success())
-            .unwrap_or(false)
-    }
-
-    fn smoke_fixture_dir() -> PathBuf {
-        PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("../../benchmarks/fixtures/scarb_smoke")
-            .canonicalize()
-            .expect("failed to locate scarb_smoke fixture")
-    }
-
-    fn copy_dir_recursive(src: &Path, dst: &Path) {
-        for entry in walkdir::WalkDir::new(src) {
-            let entry = entry.expect("failed to traverse fixture directory");
-            let rel = entry
-                .path()
-                .strip_prefix(src)
-                .expect("failed to strip fixture prefix");
-            let out = dst.join(rel);
-            if entry.file_type().is_dir() {
-                fs::create_dir_all(&out).expect("failed to create fixture output directory");
-            } else {
-                if let Some(parent) = out.parent() {
-                    fs::create_dir_all(parent).expect("failed to create fixture parent");
-                }
-                fs::copy(entry.path(), &out).expect("failed to copy fixture file");
-            }
-        }
-    }
-
-    fn prepare_smoke_workspace(prefix: &str) -> PathBuf {
-        let dir = unique_test_dir(prefix);
-        copy_dir_recursive(&smoke_fixture_dir(), &dir);
-        dir
-    }
-
-    fn smoke_common_args(manifest_path: &Path) -> BuildCommonArgs {
-        BuildCommonArgs {
-            manifest_path: Some(manifest_path.to_path_buf()),
-            package: None,
-            workspace: false,
-            features: Vec::new(),
-            offline: false,
-            release: false,
-            profile: None,
-        }
-    }
-
-    fn run_smoke_cached_build(
-        common: &BuildCommonArgs,
-        manifest_path: &Path,
-        workspace_root: &Path,
-        profile: &str,
-        session_key: &str,
-    ) -> Result<(CommandRun, bool, String, BuildPhaseTelemetry)> {
-        run_build_with_uc_cache(
-            common,
-            manifest_path,
-            workspace_root,
-            profile,
-            session_key,
-            true,
-            false,
-        )
-    }
-
-    #[test]
-    fn daemon_metadata_request_roundtrip_preserves_fields() {
-        let args = MetadataArgs {
-            manifest_path: Some(PathBuf::from("/tmp/workspace/Scarb.toml")),
-            format_version: 2,
-            daemon_mode: DaemonModeArg::Auto,
-            offline: true,
-            global_cache_dir: Some(PathBuf::from("/tmp/scarb-cache")),
-            report_path: None,
-        };
-        let request =
-            daemon_metadata_request_from_args(&args, Path::new("/tmp/workspace/Scarb.toml"), true);
-        let restored = metadata_args_from_daemon_request(&request);
-
-        assert_eq!(
-            restored
-                .manifest_path
-                .as_ref()
-                .expect("manifest path missing"),
-            Path::new("/tmp/workspace/Scarb.toml")
-        );
-        assert_eq!(restored.format_version, 2);
-        assert_eq!(request.protocol_version, DAEMON_PROTOCOL_VERSION);
-        assert!(restored.offline);
-        assert_eq!(
-            restored.global_cache_dir,
-            Some(PathBuf::from("/tmp/scarb-cache"))
-        );
-        assert_eq!(restored.daemon_mode as u8, DaemonModeArg::Off as u8);
-        assert!(request.capture_output);
-    }
-
-    #[test]
-    fn daemon_build_request_roundtrip_preserves_async_cache_persist() {
-        let common = BuildCommonArgs {
-            manifest_path: Some(PathBuf::from("/tmp/workspace/Scarb.toml")),
-            package: Some("demo".to_string()),
-            workspace: true,
-            features: vec!["feature_a".to_string(), "feature_b".to_string()],
-            offline: true,
-            release: false,
-            profile: Some("dev".to_string()),
-        };
-        let request = daemon_build_request_from_common(
-            &common,
-            Path::new("/tmp/workspace/Scarb.toml"),
-            true,
-            false,
-        );
-        let restored = common_args_from_daemon_request(&request);
-
-        assert!(request.async_cache_persist);
-        assert!(!request.capture_output);
-        assert_eq!(request.protocol_version, DAEMON_PROTOCOL_VERSION);
-        assert_eq!(restored.package, common.package);
-        assert_eq!(restored.workspace, common.workspace);
-        assert_eq!(restored.features, common.features);
-        assert_eq!(restored.offline, common.offline);
-        assert_eq!(restored.release, common.release);
-        assert_eq!(restored.profile, common.profile);
-    }
-
-    #[test]
-    fn daemon_build_request_serialization_supports_async_cache_persist_wire_field() {
-        let request = DaemonRequest::Build(DaemonBuildRequest {
-            protocol_version: DAEMON_PROTOCOL_VERSION.to_string(),
-            manifest_path: "/tmp/workspace/Scarb.toml".to_string(),
-            package: None,
-            workspace: false,
-            features: vec!["feature_a".to_string()],
-            offline: false,
-            release: false,
-            profile: None,
-            async_cache_persist: true,
-            capture_output: true,
-        });
-        let json = serde_json::to_string(&request).expect("failed to encode request");
-        assert!(json.contains("\"type\":\"build\""));
-        assert!(json.contains("\"async_cache_persist\":true"));
-        assert!(json.contains("\"capture_output\":true"));
-
-        let decoded: DaemonRequest =
-            serde_json::from_str(&json).expect("failed to decode daemon request");
-        match decoded {
-            DaemonRequest::Build(payload) => {
-                assert!(payload.async_cache_persist);
-                assert!(payload.capture_output);
-                assert_eq!(payload.protocol_version, DAEMON_PROTOCOL_VERSION);
-                assert_eq!(payload.manifest_path, "/tmp/workspace/Scarb.toml");
-                assert_eq!(payload.features, vec!["feature_a".to_string()]);
-            }
-            _ => panic!("expected build request"),
-        }
-    }
-
-    #[test]
-    fn daemon_build_request_defaults_capture_output_when_missing_from_wire() {
-        let json = format!(
-            "{{\"type\":\"build\",\"protocol_version\":\"{}\",\"manifest_path\":\"/tmp/workspace/Scarb.toml\",\"package\":null,\"workspace\":false,\"features\":[],\"offline\":false,\"release\":false,\"profile\":null,\"async_cache_persist\":false}}",
-            DAEMON_PROTOCOL_VERSION
-        );
-        let decoded: DaemonRequest =
-            serde_json::from_str(&json).expect("failed to decode daemon request");
-        match decoded {
-            DaemonRequest::Build(payload) => {
-                assert!(payload.capture_output);
-                assert_eq!(payload.protocol_version, DAEMON_PROTOCOL_VERSION);
-            }
-            _ => panic!("expected build request"),
-        }
-    }
-
-    #[test]
-    fn daemon_metadata_request_serialization_supports_wire_format() {
-        let request = DaemonRequest::Metadata(DaemonMetadataRequest {
-            protocol_version: DAEMON_PROTOCOL_VERSION.to_string(),
-            manifest_path: "/tmp/workspace/Scarb.toml".to_string(),
-            format_version: 1,
-            offline: false,
-            global_cache_dir: None,
-            capture_output: false,
-        });
-        let json = serde_json::to_string(&request).expect("failed to encode request");
-        assert!(json.contains("\"type\":\"metadata\""));
-
-        let decoded: DaemonRequest =
-            serde_json::from_str(&json).expect("failed to decode daemon request");
-        match decoded {
-            DaemonRequest::Metadata(payload) => {
-                assert_eq!(payload.protocol_version, DAEMON_PROTOCOL_VERSION);
-                assert_eq!(payload.manifest_path, "/tmp/workspace/Scarb.toml");
-                assert_eq!(payload.format_version, 1);
-                assert!(!payload.offline);
-                assert!(!payload.capture_output);
-                assert!(payload.global_cache_dir.is_none());
-            }
-            _ => panic!("expected metadata request"),
-        }
-    }
-
-    #[test]
-    fn write_uc_toml_normalizes_windows_manifest_path() {
-        let dir = unique_test_dir("uc-write-uc-toml");
-        let output = dir.join("Uc.toml");
-
-        write_uc_toml(
-            &output,
-            Path::new(r"C:\Users\foo\project\Scarb.toml"),
-            Some("demo"),
-            Some("0.1.0"),
-            Some("2024_07"),
-        )
-        .expect("failed to write Uc.toml");
-
-        let body = fs::read_to_string(&output).expect("failed to read Uc.toml");
-        assert!(body.contains(r#"scarb_manifest = "C:/Users/foo/project/Scarb.toml""#));
-        assert!(!body.contains(r#"C:\Users\foo\project\Scarb.toml"#));
-
-        fs::remove_dir_all(&dir).ok();
-    }
-
-    #[test]
-    fn scarb_metadata_command_includes_daemon_independent_flags() {
-        let args = MetadataArgs {
-            manifest_path: Some(PathBuf::from("/tmp/workspace/Scarb.toml")),
-            format_version: 2,
-            daemon_mode: DaemonModeArg::Require,
-            offline: true,
-            global_cache_dir: Some(PathBuf::from("/tmp/scarb-cache")),
-            report_path: None,
-        };
-        let (_, command_vec) =
-            scarb_metadata_command(&args, Path::new("/tmp/workspace/Scarb.toml"));
-        assert_eq!(
-            command_vec,
-            vec![
-                "scarb",
-                "--manifest-path",
-                "/tmp/workspace/Scarb.toml",
-                "--offline",
-                "--global-cache-dir",
-                "/tmp/scarb-cache",
-                "metadata",
-                "--format-version",
-                "2",
-            ]
-        );
-    }
-
-    #[test]
-    fn parse_scarb_semver_extracts_triplet() {
-        assert_eq!(
-            parse_scarb_semver("scarb 2.14.0 (682b29e13 2025-11-25)").unwrap(),
-            (2, 14, 0)
-        );
-        assert!(parse_scarb_semver("invalid-output").is_err());
-    }
-
-    #[test]
-    fn scarb_version_line_uses_env_override() {
-        let _guard = integration_env_lock()
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        std::env::set_var("UC_SCARB_VERSION_LINE", "scarb 9.9.9 (override)");
-        let version = scarb_version_line().expect("override version should be accepted");
-        assert_eq!(version, "scarb 9.9.9 (override)");
-        std::env::remove_var("UC_SCARB_VERSION_LINE");
-    }
-
-    #[test]
-    fn build_env_fingerprint_tracks_prefixed_vars_only() {
-        let _guard = integration_env_lock()
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        std::env::remove_var("UC_BUILD_ENV_PREFIXES_EXTRA");
-        std::env::remove_var("SCARB_TEST_UC_FINGERPRINT");
-        std::env::remove_var("UC_TEST_UNRELATED_FINGERPRINT");
-
-        let baseline = compute_build_env_fingerprint();
-        std::env::set_var("SCARB_TEST_UC_FINGERPRINT", "v1");
-        let with_prefixed = compute_build_env_fingerprint();
-        assert_ne!(baseline, with_prefixed);
-
-        std::env::set_var("UC_TEST_UNRELATED_FINGERPRINT", "noise");
-        let with_unrelated = compute_build_env_fingerprint();
-        assert_eq!(with_prefixed, with_unrelated);
-
-        std::env::set_var("SCARB_TEST_UC_FINGERPRINT", "v2");
-        let with_prefixed_change = compute_build_env_fingerprint();
-        assert_ne!(with_prefixed, with_prefixed_change);
-
-        std::env::remove_var("SCARB_TEST_UC_FINGERPRINT");
-        std::env::remove_var("UC_TEST_UNRELATED_FINGERPRINT");
-    }
-
-    #[test]
-    fn build_env_fingerprint_supports_extra_prefixes() {
-        let _guard = integration_env_lock()
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        std::env::remove_var("UC_BUILD_ENV_PREFIXES_EXTRA");
-        std::env::set_var("XBUILD_TEST_FINGERPRINT", "v1");
-        let without_extra = compute_build_env_fingerprint();
-
-        std::env::set_var("UC_BUILD_ENV_PREFIXES_EXTRA", "XBUILD_");
-        let with_extra = compute_build_env_fingerprint();
-        assert_ne!(without_extra, with_extra);
-
-        std::env::remove_var("XBUILD_TEST_FINGERPRINT");
-        std::env::remove_var("UC_BUILD_ENV_PREFIXES_EXTRA");
-    }
-
-    #[test]
-    fn session_input_cache_key_is_order_independent_for_features() {
-        let common_a = BuildCommonArgs {
-            manifest_path: Some(PathBuf::from("/tmp/workspace/Scarb.toml")),
-            package: Some("demo".to_string()),
-            workspace: false,
-            features: vec!["b".to_string(), "a".to_string()],
-            offline: false,
-            release: false,
-            profile: None,
-        };
-        let common_b = BuildCommonArgs {
-            manifest_path: Some(PathBuf::from("/tmp/workspace/Scarb.toml")),
-            package: Some("demo".to_string()),
-            workspace: false,
-            features: vec!["a".to_string(), "b".to_string(), "a".to_string()],
-            offline: false,
-            release: false,
-            profile: None,
-        };
-
-        let key_a = session_input_cache_key(
-            &common_a,
-            Path::new("/tmp/workspace/Scarb.toml"),
-            "dev",
-            "scarb 2.14.0",
-            "env-a",
-        );
-        let key_b = session_input_cache_key(
-            &common_b,
-            Path::new("/tmp/workspace/Scarb.toml"),
-            "dev",
-            "scarb 2.14.0",
-            "env-a",
-        );
-        assert_eq!(key_a, key_b);
-    }
-
-    #[test]
-    fn session_input_cache_key_changes_with_build_env_fingerprint() {
-        let common = BuildCommonArgs {
-            manifest_path: Some(PathBuf::from("/tmp/workspace/Scarb.toml")),
-            package: Some("demo".to_string()),
-            workspace: false,
-            features: vec!["a".to_string()],
-            offline: false,
-            release: false,
-            profile: None,
-        };
-        let key_a = session_input_cache_key(
-            &common,
-            Path::new("/tmp/workspace/Scarb.toml"),
-            "dev",
-            "scarb 2.14.0",
-            "env-a",
-        );
-        let key_b = session_input_cache_key(
-            &common,
-            Path::new("/tmp/workspace/Scarb.toml"),
-            "dev",
-            "scarb 2.14.0",
-            "env-b",
-        );
-        assert_ne!(key_a, key_b);
-    }
-
-    #[test]
-    fn cache_budget_enforcement_stride_triggers_every_nth_persist() {
-        assert!(!should_enforce_cache_size_budget_for_persist_index(1, 8));
-        assert!(!should_enforce_cache_size_budget_for_persist_index(7, 8));
-        assert!(should_enforce_cache_size_budget_for_persist_index(8, 8));
-        assert!(!should_enforce_cache_size_budget_for_persist_index(9, 8));
-        assert!(should_enforce_cache_size_budget_for_persist_index(16, 8));
-    }
-
-    #[test]
-    fn cache_budget_enforcement_stride_one_triggers_every_persist() {
-        for persist_index in 1..=8 {
-            assert!(should_enforce_cache_size_budget_for_persist_index(
-                persist_index,
-                1
-            ));
-        }
-    }
-
-    #[test]
-    fn cache_budget_enforcement_state_respects_interval_and_first_arm() {
-        assert!(!should_enforce_cache_size_budget_for_state(
-            8, 8, 10_000, 0, 60_000
-        ));
-        assert!(!should_enforce_cache_size_budget_for_state(
-            8, 8, 20_000, 10_000, 60_000
-        ));
-        assert!(should_enforce_cache_size_budget_for_state(
-            16, 8, 80_000, 10_000, 60_000
-        ));
-    }
-
-    #[test]
-    fn cache_budget_enforcement_state_interval_zero_uses_stride_only() {
-        assert!(!should_enforce_cache_size_budget_for_state(
-            7, 8, 1_000, 0, 0
-        ));
-        assert!(should_enforce_cache_size_budget_for_state(
-            8, 8, 1_000, 0, 0
-        ));
-    }
-
-    #[test]
-    fn parse_metadata_format_version_accepts_supported_values() {
-        assert_eq!(parse_metadata_format_version("1").unwrap(), 1);
-        assert_eq!(parse_metadata_format_version("2").unwrap(), 2);
-        assert!(parse_metadata_format_version("3").is_err());
-    }
-
-    #[test]
-    fn validate_daemon_protocol_version_rejects_mismatch() {
-        let err = validate_daemon_protocol_version("0.0.0")
-            .err()
-            .expect("expected mismatch");
-        assert!(
-            format!("{err:#}").contains("daemon protocol mismatch"),
-            "unexpected error: {err:#}"
-        );
-    }
-
-    #[test]
-    fn persist_artifact_object_materializes_destination() {
-        let dir = unique_test_dir("uc-persist-object");
-        let source = dir.join("source.bin");
-        let destination = dir.join("objects/aa/object.bin");
-        fs::create_dir_all(
-            destination
-                .parent()
-                .expect("destination should have parent directory"),
-        )
-        .expect("failed to create object directory");
-        fs::write(&source, b"artifact-bytes").expect("failed to write source object");
-
-        persist_artifact_object(&source, &destination).expect("persist should succeed");
-
-        let restored = fs::read(&destination).expect("failed to read destination");
-        assert_eq!(restored, b"artifact-bytes");
-        fs::remove_dir_all(&dir).ok();
-    }
-
-    #[test]
-    fn restore_cache_object_overwrites_existing_file() {
-        let dir = unique_test_dir("uc-restore-object");
-        let source = dir.join("source.bin");
-        let destination = dir.join("target/output.bin");
-        fs::create_dir_all(
-            destination
-                .parent()
-                .expect("destination should have parent directory"),
-        )
-        .expect("failed to create destination directory");
-        fs::write(&source, b"fresh-object").expect("failed to write source object");
-        fs::write(&destination, b"stale-object").expect("failed to write stale destination");
-
-        restore_cache_object(&source, &destination).expect("restore should succeed");
-
-        let restored = fs::read(&destination).expect("failed to read restored object");
-        assert_eq!(restored, b"fresh-object");
-        fs::remove_dir_all(&dir).ok();
-    }
-
-    #[test]
-    fn restore_cached_artifacts_skips_object_restore_when_index_matches() {
-        let dir = unique_test_dir("uc-restore-index-hit");
-        let workspace = dir.join("workspace");
-        let target_root = workspace.join("target/dev");
-        let cache_root = workspace.join(".uc/cache");
-        let objects_dir = cache_root.join("objects");
-        fs::create_dir_all(&target_root).expect("failed to create target root");
-        fs::create_dir_all(&objects_dir).expect("failed to create objects root");
-
-        let output = target_root.join("demo.sierra.json");
-        fs::write(&output, b"cached-artifact").expect("failed to write target artifact");
-        let output_metadata = fs::metadata(&output).expect("failed to stat target artifact");
-        let expected_hash = hash_file_blake3(&output).expect("failed to hash target artifact");
-
-        let mut artifact_index = ArtifactIndex::empty();
-        artifact_index.entries.insert(
-            "demo.sierra.json".to_string(),
-            ArtifactIndexEntry {
-                size_bytes: output_metadata.len(),
-                modified_unix_ms: metadata_modified_unix_ms(&output_metadata)
-                    .expect("failed to read target mtime"),
-                blake3_hex: expected_hash.clone(),
-            },
-        );
-        save_artifact_index(&cache_root.join("artifact-index-v1.json"), &artifact_index)
-            .expect("failed to write artifact index");
-
-        let artifacts = vec![CachedArtifact {
-            relative_path: "demo.sierra.json".to_string(),
-            blake3_hex: expected_hash,
-            size_bytes: output_metadata.len(),
-            object_rel_path: "aa/missing-object.bin".to_string(),
-        }];
-
-        let restored =
-            restore_cached_artifacts(&workspace, "dev", &cache_root, &objects_dir, &artifacts)
-                .expect("restore should succeed");
-        assert!(restored);
-        assert_eq!(
-            fs::read(&output).expect("failed to read target artifact"),
-            b"cached-artifact"
-        );
-        fs::remove_dir_all(&dir).ok();
-    }
-
-    #[test]
-    fn restore_cached_artifacts_restores_when_target_mismatch() {
-        let dir = unique_test_dir("uc-restore-index-miss");
-        let workspace = dir.join("workspace");
-        let target_root = workspace.join("target/dev");
-        let cache_root = workspace.join(".uc/cache");
-        let objects_dir = cache_root.join("objects/aa");
-        fs::create_dir_all(&target_root).expect("failed to create target root");
-        fs::create_dir_all(&objects_dir).expect("failed to create objects root");
-
-        let output = target_root.join("demo.sierra.json");
-        fs::write(&output, b"stale-artifact").expect("failed to write stale artifact");
-        let object = objects_dir.join("fresh-object.bin");
-        fs::write(&object, b"fresh-artifact").expect("failed to write cache object");
-        let expected_hash = hash_file_blake3(&object).expect("failed to hash cache object");
-        let object_metadata = fs::metadata(&object).expect("failed to stat cache object");
-
-        let artifacts = vec![CachedArtifact {
-            relative_path: "demo.sierra.json".to_string(),
-            blake3_hex: expected_hash,
-            size_bytes: object_metadata.len(),
-            object_rel_path: "aa/fresh-object.bin".to_string(),
-        }];
-
-        let restored = restore_cached_artifacts(
-            &workspace,
-            "dev",
-            &cache_root,
-            &cache_root.join("objects"),
-            &artifacts,
-        )
-        .expect("restore should succeed");
-        assert!(restored);
-        assert_eq!(
-            fs::read(&output).expect("failed to read restored artifact"),
-            b"fresh-artifact"
-        );
-        fs::remove_dir_all(&dir).ok();
-    }
-
-    #[test]
-    fn validate_manifest_dependency_sanity_rejects_self_dependency() {
-        let dir = unique_test_dir("uc-self-dependency");
-        let manifest_path = dir.join("Scarb.toml");
-        fs::write(
-            &manifest_path,
-            r#"[package]
-name = "demo"
-version = "0.1.0"
-edition = "2024_07"
-
-[dependencies]
-demo = "1.0.0"
-"#,
-        )
-        .expect("failed to write manifest");
-
-        let result = validate_manifest_dependency_sanity(&manifest_path);
-        assert!(result.is_err());
-        fs::remove_dir_all(&dir).ok();
-    }
-
-    #[test]
-    fn normalize_fingerprint_path_normalizes_windows_prefix() {
-        assert_eq!(
-            normalize_fingerprint_path(Path::new(r"\\?\C:\tmp\demo\Scarb.toml")),
-            "C:/tmp/demo/Scarb.toml"
-        );
-    }
-
-    #[test]
-    fn hot_fingerprint_reuses_digest_when_tracked_metadata_is_unchanged() {
-        let dir = unique_test_dir("uc-hot-fingerprint-hit");
-        let src_dir = dir.join("src");
-        fs::create_dir_all(&src_dir).expect("failed to create src dir");
-        let source = src_dir.join("lib.cairo");
-        fs::write(&source, b"fn main() -> felt252 { 1 }").expect("failed to write source file");
-
-        let source_metadata = fs::metadata(&source).expect("failed to stat source");
-        let src_dir_metadata = fs::metadata(&src_dir).expect("failed to stat src dir");
-        let root_metadata = fs::metadata(&dir).expect("failed to stat workspace");
-
-        let mut entries = BTreeMap::new();
-        entries.insert(
-            "src/lib.cairo".to_string(),
-            FingerprintIndexEntry {
-                size_bytes: source_metadata.len(),
-                modified_unix_ms: metadata_modified_unix_ms(&source_metadata)
-                    .expect("failed to read source mtime"),
-                blake3_hex: "abc123".to_string(),
-            },
-        );
-        let mut directories = BTreeMap::new();
-        directories.insert(
-            ".".to_string(),
-            metadata_modified_unix_ms(&root_metadata).expect("failed to read workspace mtime"),
-        );
-        directories.insert(
-            "src".to_string(),
-            metadata_modified_unix_ms(&src_dir_metadata).expect("failed to read src dir mtime"),
-        );
-
-        let index = FingerprintIndex {
-            schema_version: FINGERPRINT_INDEX_SCHEMA_VERSION,
-            entries,
-            directories,
-            context_digest: Some("ctx".to_string()),
-            last_fingerprint: Some("fp-hot-hit".to_string()),
-        };
-        let reused = try_reuse_hot_fingerprint(&dir, &index, "ctx", u64::MAX, 0)
-            .expect("hot-path check should succeed");
-        assert_eq!(reused, Some("fp-hot-hit".to_string()));
-        fs::remove_dir_all(&dir).ok();
-    }
-
-    #[test]
-    fn hot_fingerprint_invalidates_when_tracked_directory_mtime_changes() {
-        let dir = unique_test_dir("uc-hot-fingerprint-dir-change");
-        let src_dir = dir.join("src");
-        fs::create_dir_all(&src_dir).expect("failed to create src dir");
-        let source = src_dir.join("lib.cairo");
-        fs::write(&source, b"fn main() -> felt252 { 1 }").expect("failed to write source file");
-
-        let source_metadata = fs::metadata(&source).expect("failed to stat source");
-        let src_dir_metadata = fs::metadata(&src_dir).expect("failed to stat src dir");
-        let root_metadata = fs::metadata(&dir).expect("failed to stat workspace");
-
-        let mut entries = BTreeMap::new();
-        entries.insert(
-            "src/lib.cairo".to_string(),
-            FingerprintIndexEntry {
-                size_bytes: source_metadata.len(),
-                modified_unix_ms: metadata_modified_unix_ms(&source_metadata)
-                    .expect("failed to read source mtime"),
-                blake3_hex: "abc123".to_string(),
-            },
-        );
-        let mut directories = BTreeMap::new();
-        directories.insert(
-            ".".to_string(),
-            metadata_modified_unix_ms(&root_metadata).expect("failed to read workspace mtime"),
-        );
-        directories.insert(
-            "src".to_string(),
-            metadata_modified_unix_ms(&src_dir_metadata).expect("failed to read src dir mtime"),
-        );
-
-        let index = FingerprintIndex {
-            schema_version: FINGERPRINT_INDEX_SCHEMA_VERSION,
-            entries,
-            directories,
-            context_digest: Some("ctx".to_string()),
-            last_fingerprint: Some("fp-hot-hit".to_string()),
-        };
-
-        thread::sleep(Duration::from_millis(20));
-        fs::write(src_dir.join("new.cairo"), b"fn extra() -> felt252 { 2 }")
-            .expect("failed to add new source");
-
-        let reused = try_reuse_hot_fingerprint(&dir, &index, "ctx", u64::MAX, 0)
-            .expect("hot-path check should succeed");
-        assert!(reused.is_none());
-        fs::remove_dir_all(&dir).ok();
-    }
-
-    #[test]
-    fn strip_cairo_comments_preserves_literals() {
-        let source = br#"fn demo() {
-    let url = "http://localhost";
-    let marker = '//';
-    // remove this comment
-    /* and this block */
-}
-"#;
-        let stripped = strip_cairo_comments(source);
-        let text = String::from_utf8(stripped).expect("stripped output should be utf-8");
-        assert!(text.contains("http://localhost"));
-        assert!(text.contains("let marker = '//';"));
-        assert!(!text.contains("remove this comment"));
-        assert!(!text.contains("and this block"));
-    }
-
-    #[test]
-    fn fingerprint_ignores_cairo_comment_only_edits() {
-        let workspace = prepare_smoke_workspace("uc-fingerprint-comments");
-        let manifest_path = workspace.join("Scarb.toml");
-        let common = smoke_common_args(&manifest_path);
-        let profile = effective_profile(&common);
-        let lib_path = workspace.join("src/lib.cairo");
-
-        let original = compute_build_fingerprint_with_scarb_version(
-            &workspace,
-            &manifest_path,
-            &common,
-            &profile,
-            None,
-            "scarb 2.14.0 (test)",
-        )
-        .expect("failed to compute baseline fingerprint");
-
-        fs::write(
-            &lib_path,
-            format!(
-                "{}\n// comment-only change for fingerprint test\n",
-                fs::read_to_string(&lib_path).expect("failed to read lib.cairo")
-            ),
-        )
-        .expect("failed to append comment");
-
-        let with_comment = compute_build_fingerprint_with_scarb_version(
-            &workspace,
-            &manifest_path,
-            &common,
-            &profile,
-            None,
-            "scarb 2.14.0 (test)",
-        )
-        .expect("failed to compute comment fingerprint");
-        assert_eq!(original, with_comment);
-
-        let updated = fs::read_to_string(&lib_path).expect("failed to read updated lib.cairo");
-        fs::write(
-            &lib_path,
-            updated.replace(
-                "BENCH_EDIT_SEED_BIAS: felt252 = 0",
-                "BENCH_EDIT_SEED_BIAS: felt252 = 1",
-            ),
-        )
-        .expect("failed to write semantic change");
-
-        let with_semantic_change = compute_build_fingerprint_with_scarb_version(
-            &workspace,
-            &manifest_path,
-            &common,
-            &profile,
-            None,
-            "scarb 2.14.0 (test)",
-        )
-        .expect("failed to compute semantic-change fingerprint");
-        assert_ne!(original, with_semantic_change);
-
-        fs::remove_dir_all(&workspace).ok();
-    }
-
-    #[test]
-    fn async_persist_error_queue_retains_multiple_failures() {
-        let _ = take_async_persist_errors();
-        record_async_persist_error("err-a".to_string());
-        record_async_persist_error("err-b".to_string());
-
-        assert_eq!(
-            take_async_persist_errors(),
-            vec!["err-a".to_string(), "err-b".to_string()]
-        );
-    }
-
-    #[test]
-    fn async_persist_error_queue_drops_oldest_when_over_capacity() {
-        let _ = take_async_persist_errors();
-        let overflow = ASYNC_PERSIST_ERROR_QUEUE_LIMIT + 3;
-        for i in 0..overflow {
-            record_async_persist_error(format!("err-{i}"));
-        }
-
-        let drained = take_async_persist_errors();
-        assert_eq!(drained.len(), ASYNC_PERSIST_ERROR_QUEUE_LIMIT);
-        assert_eq!(drained.first().map(String::as_str), Some("err-3"));
-        assert_eq!(
-            drained.last().map(String::as_str),
-            Some(format!("err-{}", overflow - 1).as_str())
-        );
-    }
-
-    #[test]
-    fn compute_build_fingerprint_changes_when_scarb_version_changes() {
-        let _guard = integration_env_lock()
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        let workspace = prepare_smoke_workspace("uc-fingerprint-version");
-        let manifest_path = workspace.join("Scarb.toml");
-        let common = smoke_common_args(&manifest_path);
-        let profile = effective_profile(&common);
-
-        let v1 = compute_build_fingerprint_with_scarb_version(
-            &workspace,
-            &manifest_path,
-            &common,
-            &profile,
-            None,
-            "scarb 2.14.0 (test)",
-        )
-        .expect("failed to compute fingerprint for v1");
-        let v2 = compute_build_fingerprint_with_scarb_version(
-            &workspace,
-            &manifest_path,
-            &common,
-            &profile,
-            None,
-            "scarb 2.15.0 (test)",
-        )
-        .expect("failed to compute fingerprint for v2");
-        assert_ne!(v1, v2);
-
-        fs::remove_dir_all(&workspace).ok();
-    }
-
-    #[test]
-    fn run_build_with_uc_cache_hits_after_initial_compile() {
-        let _guard = integration_env_lock()
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        if !scarb_available() {
-            return;
-        }
-        let workspace = prepare_smoke_workspace("uc-cache-hit");
-        let manifest_path = workspace.join("Scarb.toml");
-        let common = smoke_common_args(&manifest_path);
-        let profile = effective_profile(&common);
-        let session_key = build_session_input(&common, &manifest_path, &profile)
-            .expect("failed to compute session input")
-            .deterministic_key_hex();
-
-        let (first_run, first_hit, first_fingerprint, _) =
-            run_smoke_cached_build(&common, &manifest_path, &workspace, &profile, &session_key)
-                .expect("first build should succeed");
-        assert_eq!(first_run.exit_code, 0);
-        assert!(!first_hit);
-
-        let (second_run, second_hit, second_fingerprint, _) =
-            run_smoke_cached_build(&common, &manifest_path, &workspace, &profile, &session_key)
-                .expect("second build should succeed");
-        assert_eq!(second_run.exit_code, 0);
-        assert!(second_hit);
-        assert_eq!(first_fingerprint, second_fingerprint);
-
-        fs::remove_dir_all(&workspace).ok();
-    }
-
-    #[test]
-    fn run_build_with_uc_cache_recovers_from_corrupted_entry() {
-        let _guard = integration_env_lock()
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        if !scarb_available() {
-            return;
-        }
-        let workspace = prepare_smoke_workspace("uc-cache-corruption");
-        let manifest_path = workspace.join("Scarb.toml");
-        let common = smoke_common_args(&manifest_path);
-        let profile = effective_profile(&common);
-        let session_key = build_session_input(&common, &manifest_path, &profile)
-            .expect("failed to compute session input")
-            .deterministic_key_hex();
-
-        run_smoke_cached_build(&common, &manifest_path, &workspace, &profile, &session_key)
-            .expect("initial build should succeed");
-
-        let entry_path = workspace
-            .join(".uc/cache/build")
-            .join(format!("{session_key}.json"));
-        fs::write(&entry_path, b"{not-json").expect("failed to corrupt cache entry");
-
-        let (run, cache_hit, _, _) =
-            run_smoke_cached_build(&common, &manifest_path, &workspace, &profile, &session_key)
-                .expect("build should recover from corrupted cache entry");
-        assert_eq!(run.exit_code, 0);
-        assert!(!cache_hit);
-
-        fs::remove_dir_all(&workspace).ok();
-    }
-
-    #[test]
-    fn run_build_with_uc_cache_allows_concurrent_builds() {
-        let _guard = integration_env_lock()
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        if !scarb_available() {
-            return;
-        }
-        let workspace = prepare_smoke_workspace("uc-cache-concurrency");
-        let manifest_path = workspace.join("Scarb.toml");
-        let common = smoke_common_args(&manifest_path);
-        let profile = effective_profile(&common);
-        let session_key = build_session_input(&common, &manifest_path, &profile)
-            .expect("failed to compute session input")
-            .deterministic_key_hex();
-
-        let workspace_a = workspace.clone();
-        let manifest_a = manifest_path.clone();
-        let profile_a = profile.clone();
-        let session_a = session_key.clone();
-        let common_a = common.clone();
-        let worker_a = thread::spawn(move || {
-            run_smoke_cached_build(&common_a, &manifest_a, &workspace_a, &profile_a, &session_a)
-        });
-
-        let workspace_b = workspace.clone();
-        let manifest_b = manifest_path.clone();
-        let profile_b = profile.clone();
-        let session_b = session_key.clone();
-        let common_b = common.clone();
-        let worker_b = thread::spawn(move || {
-            run_smoke_cached_build(&common_b, &manifest_b, &workspace_b, &profile_b, &session_b)
-        });
-
-        let (run_a, _, _, _) = worker_a
-            .join()
-            .expect("worker A panicked")
-            .expect("worker A build failed");
-        let (run_b, _, _, _) = worker_b
-            .join()
-            .expect("worker B panicked")
-            .expect("worker B build failed");
-        assert_eq!(run_a.exit_code, 0);
-        assert_eq!(run_b.exit_code, 0);
-
-        fs::remove_dir_all(&workspace).ok();
-    }
-
-    #[test]
-    fn build_session_cfg_set_changes_when_cairo_target_or_tool_changes() {
-        let dir = unique_test_dir("uc-session-cfg");
-        let manifest_path = dir.join("Scarb.toml");
-
-        fs::write(
-            &manifest_path,
-            r#"[package]
-name = "cfg_test"
-version = "0.1.0"
-edition = "2024_07"
-
-[cairo]
-allow-warnings = true
-
-[target.starknet-contract]
-sierra = true
-
-[tool.uc]
-mode = "fast"
-"#,
-        )
-        .expect("failed to write manifest");
-
-        let cfg_a = build_session_cfg_set(&manifest_path).expect("failed to compute cfg A");
-
-        fs::write(
-            &manifest_path,
-            r#"[package]
-name = "cfg_test"
-version = "0.1.0"
-edition = "2024_07"
-
-[cairo]
-allow-warnings = false
-
-[target.starknet-contract]
-sierra = true
-
-[tool.uc]
-mode = "safe"
-"#,
-        )
-        .expect("failed to rewrite manifest");
-
-        let cfg_b = build_session_cfg_set(&manifest_path).expect("failed to compute cfg B");
-        assert_ne!(cfg_a, cfg_b);
-
-        fs::remove_dir_all(&dir).ok();
-    }
-}
+mod main_tests;
